Перенес проверку доступа в отдельный миксин
