+++ api/src/apps/base/__init__.py
@@ -1,0 +1,3 @@
+from .permissions import PermissionMixin
+
+__all__ = ["PermissionMixin"]
